Backend/ETHPAY/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
bcrypt==4.2.0


